pyproject.toml
[project]
name = "tunnelo"
version = "0.1.0"
description = ""
authors = [{ name = "Kacper Wardega", email = "kwardega@stellarcyber.ai" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer (>=0.15.2,<0.16.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.406"
ruff = "^0.11.7"
ipython = "^9.4.0"

[tool.pyright]
typeCheckingMode = "strict"
reportUnnecessaryComparison = true